refactor: separate SendService.swift
