+++ p2p_wallet/Services/SendService/SendService.swift
@@ -1,46 +1,13 @@
 //
-//  SendServiceType.swift
+//  SendService.swift
 //  p2p_wallet
 //
 //  Created by Chung Tran on 27/09/2021.
 //
 
 import FeeRelayerSwift
-import Foundation
 import OrcaSwapSwift
 import RxSwift
-
-protocol SendServiceType {
-    func load() -> Completable
-    func checkAccountValidation(account: String) -> Single<Bool>
-    func isTestNet() -> Bool
-
-    func getFees(
-        from wallet: Wallet,
-        receiver: String?,
-        network: SendToken.Network,
-        payingTokenMint: String?
-    ) -> Single<SolanaSDK.FeeAmount?>
-    func getFeesInPayingToken(
-        feeInSOL: SolanaSDK.FeeAmount,
-        payingFeeWallet: Wallet
-    ) -> Single<SolanaSDK.FeeAmount?>
-
-    func getFreeTransactionFeeLimit(
-    ) -> Single<FeeRelayer.Relay.FreeTransactionFeeLimit>
-
-    func getAvailableWalletsToPayFee(
-        feeInSOL: SolanaSDK.FeeAmount
-    ) -> Single<[Wallet]>
-
-    func send(
-        from wallet: Wallet,
-        receiver: String,
-        amount: Double,
-        network: SendToken.Network,
-        payingFeeWallet: Wallet?
-    ) -> Single<String>
-}
 
 class SendService: SendServiceType {
     private let locker = NSLock()
